.tests/requirements.txt
ipykernel
nbconvert
nbformat
pytest

Pillow
dhash

geopandas
pandas
matplotlib
shapely
